CONF-2407 CONF-3135 CONF-7266
Modified the macro to be more knowledgable about field name and their various capitalizations.
Some builtin fields are mixed-case (fixVersion) and so the macro needs to know when to use
the capitalization provided by the user (e.g., custom fields) and when to use a canonical form.

Moved custom field extraction code to XmlTransformer